moves reused formatErrorMessage function to a utils folder
