+++ RoomieRanks-Frontend/src/views/LoginView.vue
@@ -21,6 +21,7 @@
 <script>
 import axios from "axios";
 import { session } from "../session.ts";
+import { formatErrorMessage } from '../utils/errorFormatter.ts';
 
 const axiosClient = axios.create({
   //NOTE: it's baseURL, not baseUrl
@@ -80,8 +81,8 @@ export default {
                   // Backend returns errors in format: { "error": "error message" }
                   const message = error.response.data?.error || "An error occurred.";
                   
-                  // Format the error message: capitalize first letter and add period if missing
-                  let errorMessage = this.formatErrorMessage(message);
+                                     // Format the error message: capitalize first letter and add period if missing
+                   let errorMessage = formatErrorMessage(message);
                   console.log("Backend error:", message);
                   
                   // Show error popup
@@ -115,27 +116,6 @@ export default {
   isInputValid() {
     return this.username   
               && this.password
-  },
-
-  formatErrorMessage(message) {
-      if (!message || typeof message !== 'string') {
-          return "An error occurred.";
-      }
-      
-      // Trim whitespace
-      let formattedMessage = message.trim();
-      
-      // Capitalize first letter
-      if (formattedMessage.length > 0) {
-          formattedMessage = formattedMessage.charAt(0).toUpperCase() + formattedMessage.slice(1);
-      }
-      
-      // Add period if it doesn't end with punctuation
-      if (formattedMessage.length > 0 && !formattedMessage.match(/[.!?]$/)) {
-          formattedMessage += '.';
-      }
-      
-      return formattedMessage;
   }
 
   }
